refactor: move createFromString and resource to CLI
