+++ packages/cli/package.json
@@ -7,6 +7,10 @@
   },
   "bugs": "https://github.com/stoplightio/prism/issues",
   "dependencies": {
+    "@stoplight/http-spec": "^2.2.3",
+    "@stoplight/json": "^3.1.2",
+    "@stoplight/json-ref-readers": "^1.1.1",
+    "@stoplight/json-ref-resolver": "^3.0.1",
     "@stoplight/prism-core": "^3.2.9",
     "@stoplight/prism-http-server": "^3.2.9",
     "chalk": "^3.0.0",
