+++ packages/http/package.json
@@ -16,10 +16,6 @@
     "node": ">=8"
   },
   "dependencies": {
-    "@stoplight/http-spec": "^2.2.3",
-    "@stoplight/json": "^3.1.2",
-    "@stoplight/json-ref-readers": "^1.1.1",
-    "@stoplight/json-ref-resolver": "^3.0.1",
     "@stoplight/prism-core": "^3.2.9",
     "@stoplight/yaml": "^3.0.2",
     "abstract-logging": "^2.0.0",
